requirements.txt
fastapi
pydantic
fastapi
uvicorn
playwright
pytesseract
pillow
beautifulsoup4
SQLAlchemy